+++ script/run-publish-test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+dart format . && dart analyze && dart doc && dart test && dart pub publish --dry-run